+++ src/imdbpie/constants.py
@@ -6,5 +6,4 @@ HOST = 'api.imdbws.com'
 BASE_URI = 'https://{}'.format(HOST)
 SEARCH_BASE_URI = 'https://v2.sg.media-imdb.com'
 USER_AGENT = 'IMDb/8.3.1 (iPhone9,4; iOS 11.2.1)'
-
 APP_KEY = '76a6cc20-6073-4290-8a2c-951b4580ae4a'