+++ HISTORY.rst
@@ -9,7 +9,8 @@ Release History
 - Fixes client to work with new API.
 - Renames most of methods on ``Imdb`` class.
 - Changes all methods on ``Imdb`` to return raw JSON resource dictionary rather than Python objects.
-- Removes some params from ``Imdb`` init method (user_agent, proxy_uri, verify_ssl, cache, anonymize).
+- Removes params from ``Imdb`` ``__init__`` method (user_agent, proxy_uri, verify_ssl, api_key, cache, anonymize).
+- Adds ``clear_cached_credentials`` method to ``Imdb`` class.
 
 
 4.4.2 (2018-01-03)